Fix crashing when no WAILA installation.

+++ src/main/java/org/jackhuang/watercraft/common/tileentity/TileEntityStandardWaterMachine.java
@@ -225,7 +225,7 @@ public abstract class TileEntityStandardWaterMachine extends
 		return null;
 	}
 
-	public double getEnergy() {
+	public int getEnergy() {
 		return this.getTankAmount();
 	}
